Move authenticatable processing to module
